Add driver and process library for the project
Add main components of driver and process for the project.
Add example input data and output data.
For driver, an exmaple usage is:
python driver.py --input_file=merged_gene_counts.txt
For process.py, it is a library that contains all helper functions
for driver.py
More techinal detail is avaible in the README.md